moves config into folder
